refactor: tests to use apollo client
