+++ tests/authentication.spec.ts
@@ -1,124 +1,109 @@
-import {
-    sendGraphQLQuery,
-} from "./test_config";
+import { gql } from "@apollo/client/core";
+import { request } from "../config";
 
-describe("Authentication should function correctly for users", () => {
-    it("should log in with valid admin credentials", async () => {
-        // Login
-        const responseLogin = await sendGraphQLQuery(
-          `mutation {
+describe("authentication", () => {
+  describe("when a user is not logged in", () => {
+    it("should return a null result for the GraphQL 'me' query", async () => {
+      const client = request.createApolloClient();
+      const query = gql`
+        query {
+          me {
+            id
+          }
+        }
+      `;
+      const response = await client.query({ query });
+      expect(response.error).toBeUndefined();
+      expect(response.data.me).toBeNull();
+    });
+  });
+
+  describe("when a user is logged in", () => {
+    describe("when the user is role ADMIN", () => {
+      it("logins in as ADMIN and returns the role from the me query", async () => {
+        const client = request.createApolloClient();
+        const mutation = gql`
+          mutation {
             login(username: "testadmin", password: "testpassword") {
-                success
-                token
-                message
+              message
+              token
+              success
             }
-          }`
-        )
+          }
+        `;
 
-        const responseJSONLogin = await responseLogin.json();
-        const responseHeadersLogin = await responseLogin.headers;
-        const sessionCookie = await responseHeadersLogin.get("set-cookie") || "";
-        expect(responseJSONLogin.data.login.success).toEqual(true);
-        expect(responseJSONLogin.data.login.token).toEqual("");
-        expect(responseJSONLogin.data.login.message).toEqual("Success");
+        const loginResponse = await client.mutate({ mutation });
+        expect(loginResponse.errors).toBeUndefined();
+        expect(loginResponse.data.login.success).toBeTruthy();
+        expect(loginResponse.data.login.message).toBeDefined();
+        expect(loginResponse.data.login.token).toBeDefined();
 
-        // Execute me query and return current user
-        const responseMeQuery = await sendGraphQLQuery(
-          `query {
+        const query = gql`
+          query {
             me {
-              id
+              role
             }
-          }`,
-          sessionCookie
-        )
-        const responseMeQueryJSON = await responseMeQuery.json();
-        expect(responseMeQueryJSON.data.me.id).toEqual(1);
-        console.log(responseMeQueryJSON);
+          }
+        `;
 
-        // Logout
-        const responseLogout = await sendGraphQLQuery(
-          `mutation {
-            logout {
-                success
-                token
-                message
-            }
-          }`,
-          sessionCookie
-        )
-
-        const responseJSONLogout = await responseLogout.json();      
-        expect(responseJSONLogout.data.logout.success).toEqual(true);
-        expect(responseJSONLogout.data.logout.token).toEqual("");
-        expect(responseJSONLogout.data.logout.message).toEqual("Logged out");
+        const response = await client.query({ query });
+        expect(response.error).toBeUndefined();
+        expect(response.data.me).not.toBeNull();
+        expect(response.data.me.role).toBe("ADMIN");
+      });
     });
-
-    it("should log in with valid user credentials", async () => {
-        // Login
-        const responseLogin = await sendGraphQLQuery(
-          `mutation {
+    describe("when the user is role USER", () => {
+      it("logins in as ADMIN and returns the role from the me query", async () => {
+        const client = request.createApolloClient();
+        const mutation = gql`
+          mutation {
             login(username: "testuser", password: "testpassword") {
-                success
-                token
-                message
+              message
+              token
+              success
             }
-          }`
-        )
+          }
+        `;
 
-        const responseJSONLogin = await responseLogin.json();
-        const responseHeadersLogin = await responseLogin.headers;
-        const sessionCookie = await responseHeadersLogin.get("set-cookie") || "";
-        expect(responseJSONLogin.data.login.success).toEqual(true);
-        expect(responseJSONLogin.data.login.token).toEqual("");
-        expect(responseJSONLogin.data.login.message).toEqual("Success");
+        const loginResponse = await client.mutate({ mutation });
+        expect(loginResponse.errors).toBeUndefined();
+        expect(loginResponse.data.login.success).toBeTruthy();
+        expect(loginResponse.data.login.message).toBeDefined();
+        expect(loginResponse.data.login.token).toBeDefined();
 
-        // Execute me query and return current user
-        const responseMeQuery = await sendGraphQLQuery(
-          `query {
+        const query = gql`
+          query {
             me {
-              id
+              role
             }
-          }`,
-          sessionCookie
-        )
-        const responseMeQueryJSON = await responseMeQuery.json();
-        expect(responseMeQueryJSON.data.me.id).toEqual(2);
-        console.log(responseMeQueryJSON);
+          }
+        `;
 
-        // Logout
-        const responseLogout = await sendGraphQLQuery(
-          `mutation {
-            logout {
-                success
-                token
-                message
-            }
-          }`,
-          sessionCookie
-        )
-
-        const responseJSONLogout = await responseLogout.json();      
-        expect(responseJSONLogout.data.logout.success).toEqual(true);
-        expect(responseJSONLogout.data.logout.token).toEqual("");
-        expect(responseJSONLogout.data.logout.message).toEqual("Logged out");
+        const response = await client.query({ query });
+        expect(response.error).toBeUndefined();
+        expect(response.data.me).not.toBeNull();
+        expect(response.data.me.role).toBe("USER");
+      });
     });
+  });
 
-    it("should fail authentication with invalid user credentials", async () => {
-        // Attempt Login
-        const responseLogin = await sendGraphQLQuery(
-          `mutation {
-            login(username: "fakeuser", password: "fakepassword") {
-                success
-                token
-                message
-            }
-          }`
-        )
+  it("should fail login with invalid credentials", async () => {
+    const client = request.createApolloClient();
+    // user doesn't exist
+    const mutation = gql`
+      mutation {
+        login(username: "test", password: "test") {
+          message
+          token
+          success
+        }
+      }
+    `;
 
-        // Expect login to fail
-        const responseJSONLogin = await responseLogin.json();
-        expect(responseJSONLogin.data.login.success).toEqual(false);
-        expect(responseJSONLogin.data.login.token).toEqual("");
-        expect(responseJSONLogin.data.login.message).toEqual("Incorrect username or password");
-    });
+    const response = await client.mutate({ mutation });
+    expect(response.errors).toBeUndefined();
+    expect(response.data.login.success).toBeFalsy();
+    expect(response.data.login.message).toBeDefined();
+    expect(response.data.login.token).toBeDefined();
+  });
 });
